Semantic search

+++ core/src/main/java/com/springproject/core/Services/VectorService.java
@@ -3,5 +3,5 @@ package com.springproject.core.Services;
 import java.util.List;
 
 public interface VectorService {
-  List<Double> getVector(String inputString);
+  List<Float> getVector(String inputString);
 }
